refactor(desktop): one-component-per-file for CheckResourcesHotkeyMount, register remaining searchable sections
Code review on the previous fix commit caught two issues:

- CheckResourcesHotkeyMount lived in CommandPaletteHost.tsx, which already
  held two other components — extracts the shared hotkey/menu-subscription
  logic to commandPalette/hooks/useCheckResourcesHotkey (used by both
  CommandPaletteTrigger and the new mount) and moves the mount itself to its
  own commandPalette/CheckResourcesHotkeyMount folder, per this repo's
  one-component-per-file / one-folder-per-component convention.
- SECTION_PATHS (consolidated from the old two-function lookup) still
  omitted browser, agents, billing, apikeys, and security — on those five
  settings pages, getSectionFromPath() returned null, so the search
  auto-redirect effect silently no-opped instead of navigating to a
  matching section. Registers all five with their real routes in both
  SECTION_PATHS and SECTION_ORDER.
