+++ apps/desktop/src/renderer/commandPalette/CheckResourcesHotkeyMount/index.ts
@@ -1,0 +1,1 @@
+export { CheckResourcesHotkeyMount } from "./CheckResourcesHotkeyMount";
