+++ apps/desktop/src/renderer/commandPalette/CommandPaletteHost.tsx
@@ -1,9 +1,8 @@
-import { useNavigate } from "@tanstack/react-router";
 import { type ReactNode, useEffect } from "react";
 import { useHotkey } from "renderer/hotkeys";
-import { electronTrpc } from "renderer/lib/electron-trpc";
 import { CommandContextProvider } from "./core/ContextProvider";
 import { useFrameStackStore } from "./core/frames";
+import { useCheckResourcesHotkey } from "./hooks/useCheckResourcesHotkey";
 import { registerAllModules } from "./modules";
 import { CommandPalette } from "./ui/CommandPalette/CommandPalette";
 import { DeleteWorkspaceMount } from "./ui/DeleteWorkspaceMount/DeleteWorkspaceMount";
@@ -32,31 +31,6 @@ export function CommandPaletteHost({ children }: { children?: ReactNode }) {
 	);
 }
 
-/**
- * Keeps CHECK_RESOURCES on the renderer's own hotkey binding (rather than a
- * native menu accelerator) so it stays user-customizable/disable-able via
- * Settings > Keyboard — see main/lib/menu.ts for why the "Resources" menu
- * item has no accelerator. The native "Resources" menu item's click opens
- * the same view.
- */
-function useCheckResourcesHotkey(onBeforeNavigate?: () => void) {
-	const navigate = useNavigate();
-
-	const openResources = () => {
-		onBeforeNavigate?.();
-		void navigate({ to: "/settings/usage/resources" });
-	};
-
-	useHotkey("CHECK_RESOURCES", openResources);
-
-	electronTrpc.menu.subscribe.useSubscription(undefined, {
-		onData: (event) => {
-			if (event.type !== "check-resources") return;
-			openResources();
-		},
-	});
-}
-
 function CommandPaletteTrigger() {
 	const setOpen = useFrameStackStore((s) => s.setOpen);
 	const reset = useFrameStackStore((s) => s.reset);
@@ -67,17 +41,5 @@ function CommandPaletteTrigger() {
 		reset();
 	});
 
-	return null;
-}
-
-/**
- * CommandPaletteHost only mounts inside the _dashboard route tree, so routes
- * outside it (Settings) need their own mount for CHECK_RESOURCES — otherwise
- * the hotkey and the native "Resources" menu item go dead the moment the
- * user navigates into Settings, including on the Usage/Resources page they
- * point at.
- */
-export function CheckResourcesHotkeyMount() {
-	useCheckResourcesHotkey();
 	return null;
 }
