+++ apps/desktop/src/renderer/commandPalette/hooks/useCheckResourcesHotkey/index.ts
@@ -1,0 +1,1 @@
+export { useCheckResourcesHotkey } from "./useCheckResourcesHotkey";
