+++ apps/desktop/src/renderer/routes/_authenticated/settings/layout.tsx
@@ -38,8 +38,10 @@ const SECTION_ORDER: SettingsSection[] = [
 	"keyboard",
 	"behavior",
 	"git",
+	"agents",
 	"terminal",
 	"links",
+	"browser",
 	"models",
 	"organization",
 	"teams",
@@ -47,6 +49,7 @@ const SECTION_ORDER: SettingsSection[] = [
 	"integrations",
 	"billing",
 	"apikeys",
+	"security",
 	"permissions",
 	"hosts",
 	"experimental",
@@ -67,11 +70,16 @@ const SECTION_PATHS: Partial<Record<SettingsSection, string>> = {
 	keyboard: "/settings/keyboard",
 	behavior: "/settings/behavior",
 	git: "/settings/git",
+	agents: "/settings/agents",
 	terminal: "/settings/terminal",
 	links: "/settings/links",
+	browser: "/settings/browser",
 	models: "/settings/models",
 	experimental: "/settings/experimental",
 	integrations: "/settings/integrations",
+	billing: "/settings/billing",
+	apikeys: "/settings/api-keys",
+	security: "/settings/security",
 	permissions: "/settings/permissions",
 	hosts: "/settings/hosts",
 	project: "/settings/projects",
